# feat: SQLite FTS5拡張サポート実装による全文検索性能向上