--- a/internal/memory/sqlite_store_test.go
+++ b/internal/memory/sqlite_store_test.go
@@ -381,3 +381,182 @@ func TestSQLiteMemoryStore_ConcurrentAccess(t *testing.T) {
 		t.Errorf("Expected 10 memories after concurrent writes, got %d", len(memories))
 	}
 }
+
+// TestSQLiteMemoryStore_FTS5Support tests FTS5 functionality specifically
+func TestSQLiteMemoryStore_FTS5Support(t *testing.T) {
+	tempDir := t.TempDir()
+	dbPath := filepath.Join(tempDir, "test_fts5.db")
+	
+	store, err := NewSQLiteMemoryStore(dbPath)
+	if err != nil {
+		t.Fatalf("Failed to create SQLite store: %v", err)
+	}
+	defer store.Close()
+	
+	// Add test memories with Japanese content
+	memories := []*Memory{
+		{
+			ID:       "test-dog-1",
+			Category: "pet",
+			Key:      "my_dog",
+			Value:    "私の犬はとてもかわいいです。毎日散歩しています。",
+			Tags:     []string{"動物", "ペット"},
+		},
+		{
+			ID:       "test-cat-1", 
+			Category: "pet",
+			Key:      "my_cat",
+			Value:    "猫も飼っています。とても優雅な動物です。",
+			Tags:     []string{"動物", "ペット"},
+		},
+		{
+			ID:       "test-bird-1",
+			Category: "nature",
+			Key:      "birds",
+			Value:    "公園で鳥を観察するのが趣味です。",
+			Tags:     []string{"動物", "自然"},
+		},
+	}
+	
+	// Save memories
+	for _, mem := range memories {
+		_, err := store.Save(mem)
+		if err != nil {
+			t.Fatalf("Failed to save memory %s: %v", mem.ID, err)
+		}
+	}
+	
+	// Test Japanese FTS5 search
+	testCases := []struct {
+		name          string
+		query         string
+		expectedCount int
+		shouldContain []string
+	}{
+		{
+			name:          "Search for 動物 (animals)",
+			query:         "動物",
+			expectedCount: 3,
+			shouldContain: []string{"test-dog-1", "test-cat-1", "test-bird-1"},
+		},
+		{
+			name:          "Search for 犬 (dog)",
+			query:         "犬",
+			expectedCount: 1,
+			shouldContain: []string{"test-dog-1"},
+		},
+		{
+			name:          "Search for ペット (pet)",
+			query:         "ペット",
+			expectedCount: 2,
+			shouldContain: []string{"test-dog-1", "test-cat-1"},
+		},
+		{
+			name:          "Search for 散歩 (walk)",
+			query:         "散歩",
+			expectedCount: 1,
+			shouldContain: []string{"test-dog-1"},
+		},
+	}
+	
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			searchQuery := SearchQuery{
+				Query:    tc.query,
+				Category: "",
+			}
+			
+			results, err := store.Search(searchQuery)
+			if err != nil {
+				t.Fatalf("Search failed: %v", err)
+			}
+			
+			if len(results) != tc.expectedCount {
+				t.Errorf("Expected %d results, got %d", tc.expectedCount, len(results))
+			}
+			
+			// Check that expected memories are included
+			resultIDs := make(map[string]bool)
+			for _, result := range results {
+				resultIDs[result.Memory.ID] = true
+			}
+			
+			for _, expectedID := range tc.shouldContain {
+				if !resultIDs[expectedID] {
+					t.Errorf("Expected result %s not found in search results", expectedID)
+				}
+			}
+		})
+	}
+}
+
+// TestSQLiteMemoryStore_FTS5Performance tests search performance with larger dataset
+func TestSQLiteMemoryStore_FTS5Performance(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping performance test in short mode")
+	}
+	
+	tempDir := t.TempDir()
+	dbPath := filepath.Join(tempDir, "test_fts5_perf.db")
+	
+	store, err := NewSQLiteMemoryStore(dbPath)
+	if err != nil {
+		t.Fatalf("Failed to create SQLite store: %v", err)
+	}
+	defer store.Close()
+	
+	// Create a larger dataset
+	numMemories := 1000
+	for i := 0; i < numMemories; i++ {
+		memory := &Memory{
+			ID:       fmt.Sprintf("perf-test-%d", i),
+			Category: "test",
+			Key:      fmt.Sprintf("test_key_%d", i),
+			Value:    fmt.Sprintf("テストデータ %d: これは性能テスト用のデータです。検索機能をテストしています。", i),
+			Tags:     []string{"テスト", "性能"},
+		}
+		
+		// Add some variety
+		if i%10 == 0 {
+			memory.Value += " 特別なキーワード"
+			memory.Tags = append(memory.Tags, "特別")
+		}
+		
+		_, err := store.Save(memory)
+		if err != nil {
+			t.Fatalf("Failed to save memory %d: %v", i, err)
+		}
+	}
+	
+	// Measure search performance
+	start := time.Now()
+	
+	searchQuery := SearchQuery{
+		Query:    "特別",
+		Category: "",
+	}
+	
+	results, err := store.Search(searchQuery)
+	if err != nil {
+		t.Fatalf("Search failed: %v", err)
+	}
+	
+	duration := time.Since(start)
+	
+	// Verify results (FTS5 has LIMIT 50, so expect min(100, 50) = 50)
+	expectedResults := numMemories / 10 // Every 10th item has "特別" = 100
+	maxResults := 50                    // FTS5 query has LIMIT 50
+	if expectedResults > maxResults {
+		expectedResults = maxResults
+	}
+	if len(results) != expectedResults {
+		t.Errorf("Expected %d results, got %d", expectedResults, len(results))
+	}
+	
+	// Performance assertion (should be fast with FTS5)
+	if duration > time.Second {
+		t.Errorf("Search took too long: %v (expected < 1s)", duration)
+	}
+	
+	t.Logf("FTS5 search of %d records took %v", numMemories, duration)
+}